Place the APIs used exclusively for workflows into WorkflowResource and workflow-persist.service.ts
