+++ core/amber/src/main/scala/edu/uci/ics/texera/web/resource/dashboard/user/workflow/WorkflowResource.scala
@@ -604,4 +604,84 @@ class WorkflowResource extends LazyLogging {
       "Private"
     }
   }
+
+  @GET
+  @Path("/owner_user")
+  def getOwnerUser(@QueryParam("wid") wid: UInteger): User = {
+    context
+      .select(
+        USER.UID,
+        USER.NAME,
+        USER.EMAIL,
+        USER.PASSWORD,
+        USER.GOOGLE_ID,
+        USER.ROLE,
+        USER.GOOGLE_AVATAR
+      )
+      .from(WORKFLOW_OF_USER)
+      .join(USER)
+      .on(WORKFLOW_OF_USER.UID.eq(USER.UID))
+      .where(WORKFLOW_OF_USER.WID.eq(wid))
+      .fetchOneInto(classOf[User])
+  }
+
+  @GET
+  @Path("/workflow_name")
+  def getWorkflowName(@QueryParam("wid") wid: UInteger): String = {
+    context
+      .select(
+        WORKFLOW.NAME
+      )
+      .from(WORKFLOW)
+      .where(WORKFLOW.WID.eq(wid))
+      .fetchOneInto(classOf[String])
+  }
+
+  @GET
+  @Path("/public/{wid}")
+  def retrievePublicWorkflow(
+      @PathParam("wid") wid: UInteger
+  ): WorkflowWithPrivilege = {
+    val workflow = workflowDao.ctx
+      .selectFrom(WORKFLOW)
+      .where(WORKFLOW.WID.eq(wid))
+      .and(WORKFLOW.IS_PUBLIC.isTrue)
+      .fetchOne()
+    WorkflowWithPrivilege(
+      workflow.getName,
+      workflow.getDescription,
+      workflow.getWid,
+      workflow.getContent,
+      workflow.getCreationTime,
+      workflow.getLastModifiedTime,
+      workflow.getIsPublic,
+      readonly = true
+    )
+  }
+
+  @GET
+  @Path("/workflow_description")
+  def getWorkflowDescription(@QueryParam("wid") wid: UInteger): String = {
+    context
+      .select(
+        WORKFLOW.DESCRIPTION
+      )
+      .from(WORKFLOW)
+      .where(WORKFLOW.WID.eq(wid))
+      .fetchOneInto(classOf[String])
+  }
+
+  @GET
+  @Path("/workflow_user_access")
+  def workflowUserAccess(
+      @QueryParam("wid") wid: UInteger
+  ): util.List[UInteger] = {
+    val records = context
+      .select(WORKFLOW_USER_ACCESS.UID)
+      .from(WORKFLOW_USER_ACCESS)
+      .where(WORKFLOW_USER_ACCESS.WID.eq(wid))
+      .fetch()
+
+    records.getValues(WORKFLOW_USER_ACCESS.UID)
+  }
 }
