+++ core/gui/src/app/hub/service/workflow/hub-workflow.service.ts
@@ -1,12 +1,7 @@
-import { HttpClient, HttpParams } from "@angular/common/http";
+import { HttpClient } from "@angular/common/http";
 import { Injectable } from "@angular/core";
 import { Observable } from "rxjs";
 import { AppSettings } from "../../../common/app-setting";
-import { HubWorkflow } from "../../component/type/hub-workflow.interface";
-import { User } from "src/app/common/type/user";
-import { Workflow } from "../../../common/type/workflow";
-import { filter, map } from "rxjs/operators";
-import { WorkflowUtilService } from "../../../workspace/service/workflow-graph/util/workflow-util.service";
 import { DashboardWorkflow } from "../../../dashboard/type/dashboard-workflow.interface";
 
 export const WORKFLOW_BASE_URL = `${AppSettings.getApiEndpoint()}/workflow`;
@@ -21,32 +16,6 @@ export class HubWorkflowService {
 
   public getWorkflowCount(): Observable<number> {
     return this.http.get<number>(`${this.BASE_URL}/count`);
-  }
-
-  public getWorkflowList(): Observable<HubWorkflow[]> {
-    return this.http.get<HubWorkflow[]>(`${this.BASE_URL}/list`);
-  }
-
-  public getOwnerUser(wid: number): Observable<User> {
-    const params = new HttpParams().set("wid", wid);
-    return this.http.get<User>(`${this.BASE_URL}/owner_user/`, { params });
-  }
-
-  public getWorkflowName(wid: number): Observable<string> {
-    const params = new HttpParams().set("wid", wid);
-    return this.http.get(`${this.BASE_URL}/workflow_name/`, { params, responseType: "text" });
-  }
-
-  public retrievePublicWorkflow(wid: number): Observable<Workflow> {
-    return this.http.get<Workflow>(`${this.BASE_URL}/public/${wid}`).pipe(
-      filter((workflow: Workflow) => workflow != null),
-      map(WorkflowUtilService.parseWorkflowInfo)
-    );
-  }
-
-  public getWorkflowDescription(wid: number): Observable<string> {
-    const params = new HttpParams().set("wid", wid);
-    return this.http.get(`${this.BASE_URL}/workflow_description/`, { params, responseType: "text" });
   }
 
   public cloneWorkflow(wid: number): Observable<number> {
@@ -89,9 +58,5 @@ export class HubWorkflowService {
 
   public getViewCount(wid: number): Observable<number> {
     return this.http.get<number>(`${this.BASE_URL}/viewCount/${wid}`);
-  }
-
-  public getWorkflowOwners(wid: number): Observable<number[]> {
-    return this.http.get<number[]>(`${this.BASE_URL}/workflowUserAccess?wid=${wid}`);
   }
 }
